Refactor JS client to allow new types (json), add outfileType option
A new outfileType option is added to the graphql-ruby-client that allows
you to specify the generated file type (i.e., for apollo-js or json). It
would be possible to add new types and their generators in the future.

The outfileType and outFile itself are independent. For example, you
could have a json file output in a .js file. It is left up to the user
to make sure they match up.

The new json outfileType is useful for non JS clients (i.e., android and
iOS).

Added tests around the new changes.

Refactored the keyValuePairs generation in /sync/generateClient.js to
be valid json (i.e., no trailing ',').