Improve RescueMiddleware error matching
Support error matching by superclasses in rescue_table.

RescueMiddleware, when attempting to rescue the errors was only matching
against the exact class if it was present in the `rescue_table`. This PR
introduces the change where we can `rescue_from` a superclass and
RescueMiddleware will be able to handle all its descendants by checking
if either the error class provided or any of it's superclasses is in the
`rescue_table`.