Add more context to input validation error message
When a GraphQL mutation gets invalid input, we raise an error with a message like

```
Variable input of type <InputType>! was provided invalid value
```

Reproducing these issues is not always obvious because which field was
invalid is not clear.

This change updates the error message to include the path to the invalid
field when it's present. The message now looks like:

```
Variable input of type <InputType>! was provided invalid value for field foo
```

Or

```
Variable input of type <InputType>! was provided invalid value for field 0.foo
```

if the field was in an array field.

There's likely some refactorings to be done here so I'm very open to
feedback. Thank you!