Fix build_from_definition for built-in types
If an SDL document contained a built-in type definition
`build_from_definition` would raise an error due to type conflicts.

This defers merging in the `BUILT_IN_TYPES` and reverse merges them to
keep any previously defined ones in the provided document.

If an existing type is late bound then the built-in type is preferred
and overrides it.