Platform Tracing: fallback transaction name
Enable a developer to provide the platform trace API a fallback name to
use when a query name is not provided.

Note that this is a *fallback*, the query name is preferred when
available.

Testing is limited to Datadog and New Relic.

Support is within any tracer that delegates to
`PlatformTracing#transaction_name`, which appears to include AppSignal
and Scout. Not sure if that is exhaustive.