Replace Ragel tokenizer with String Scanner
This commit replaces the Ragel tokenizer with a "hand written" string
scanner based tokenizer.  This tokenizer is about ~2x faster than the
previous tokenizer.

To test, I used this benchmark:

  https://gist.github.com/tenderlove/fdce1610a3eecd8426f5845391a6c373

On this branch:

```
$ ruby -I.:lib:spec bench.rb
Warming up --------------------------------------
             graphql   506.000  i/100ms
Calculating -------------------------------------
             graphql      5.090k (± 0.4%) i/s -     25.806k in   5.070096s
```

On master:

```
$ ruby -I.:lib:spec bench.rb
Warming up --------------------------------------
             graphql   284.000  i/100ms
Calculating -------------------------------------
             graphql      2.880k (± 0.4%) i/s -     14.484k in   5.029007s
```

To ensure compatibility I ran this tokenizer in parallel with the
existing tokenizer and compared the results.  There is one difference that I
think is acceptable and that is when the tokenizer encounters unknown
input.

For example:

```ruby
require "graphql/language/lexer"
require "graphql/language/token"

str = "😘"
p GraphQL::Language::Lexer.tokenize(str)
```

Output on master is this:

```
$ ruby -I.:lib:spec test.rb
[(UNKNOWN_CHAR "\xF0" [1:1]), (UNKNOWN_CHAR "\x9F" [1:2]), (UNKNOWN_CHAR "\x98" [1:3]), (UNKNOWN_CHAR "\x98" [1:4])]
```

Output on my branch is this:

```
$ ruby -I.:lib:spec test.rb
[(UNKNOWN_CHAR "😘" [1:1])]
```

This is because Ragel only understands byte streams where Ruby knows we
have a UTF-8 string.

I didn't make many changes to the helper functions that were originally
included in the Ragel parser.  Now that we're not working with byte
arrays, I think could probably improve the code in those helper
functions too.